+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All changes to this project will be documented here.
 
+## [1.2.0] - 2025-04-29
+- Feature: Vendor -> Product Categories
+- Feature: Vendor -> Product Tags
+- Feature: Vendor -> Product Attributes
+- Feature: Vendor -> Products
+- Feature: Vendor -> Product Variations
+- Feature: Admin -> Product Categories (global view)
+- Feature: Admin -> Product Tags (global view)
+- Change: Add unique email check when registering users (admin, organization, vendor, customer)
+- Change: Add email sync to default user when changing email address (organization, vendor, customer)
+
 ## [1.1.0] - 2025-03-03
 - Feature: Admin -> Audit Logs
 - Feature: Admin -> Vendors & Vendor user registration